feat: add subject api
