+++ routes/private/routes/subject.ts
@@ -1,0 +1,55 @@
+import { Type as t } from '@sinclair/typebox';
+
+import { db, op } from '@app/drizzle/db.ts';
+import * as schema from '@app/drizzle/schema';
+import { NotFoundError } from '@app/lib/error.ts';
+import { Security, Tag } from '@app/lib/openapi/index.ts';
+import * as convert from '@app/lib/types/convert.ts';
+import * as res from '@app/lib/types/res.ts';
+import { formatErrors } from '@app/lib/types/res.ts';
+import type { App } from '@app/routes/type.ts';
+
+// eslint-disable-next-line @typescript-eslint/require-await
+export async function setup(app: App) {
+  app.get(
+    '/subjects/:subjectID',
+    {
+      schema: {
+        description: '获取条目',
+        operationId: 'getSubject',
+        tags: [Tag.Subject],
+        security: [{ [Security.CookiesSession]: [], [Security.HTTPBearer]: [] }],
+        params: t.Object({
+          subjectID: t.Integer(),
+        }),
+        response: {
+          200: t.Ref(res.Subject),
+          404: t.Ref(res.Error, {
+            'x-examples': formatErrors(new NotFoundError('subject')),
+          }),
+        },
+      },
+    },
+    async ({ auth, params: { subjectID } }) => {
+      const data = await db
+        .select()
+        .from(schema.chiiSubjects)
+        .innerJoin(
+          schema.chiiSubjectFields,
+          op.eq(schema.chiiSubjects.id, schema.chiiSubjectFields.id),
+        )
+        .where(
+          op.and(
+            op.eq(schema.chiiSubjects.id, subjectID),
+            op.eq(schema.chiiSubjects.ban, 0),
+            auth.allowNsfw ? undefined : op.eq(schema.chiiSubjects.nsfw, false),
+          ),
+        )
+        .execute();
+      for (const d of data) {
+        return convert.toSubject(d.subject, d.subject_field);
+      }
+      throw new NotFoundError(`subject ${subjectID}`);
+    },
+  );
+}
